Move all auxiliary functions to a separate file, to improve code readability
